Use /dev/random by default if available.
General security principle,  always default to the most secure option. If someone wants to -s /dev/zero the still can of course,  but we assume they know what they are doing!

Not all non windows platforms have a /dev/random, you can disable it on Linux if you really want to, so we check first.